+++ src/main/java/com/rocker/services/PlantService.java
@@ -8,25 +8,25 @@ import com.rocker.model.Plant;
 
 public interface PlantService {
 
-	 public Plant addPlant(Plant plant) throws PlantException;
+	public Plant addPlant(Plant plant) throws PlantException;
 
-	    public Plant updatePlant(Plant plant) throws PlantException;
+	public Plant updatePlant(Plant plant) throws PlantException;
 
-	    public Plant deletePlantById(Integer plantId) throws PlantException;
+	public Plant deletePlantById(Integer plantId) throws PlantException;
 
-	    public Plant viewPlantById(Integer plantId) throws PlantException;
+	public Plant viewPlantById(Integer plantId) throws PlantException;
 
-	    public List<Plant> viewPlantsByName(String name) throws PlantException;
+	public List<Plant> viewPlantsByName(String name) throws PlantException;
 
-	    public List<Plant> viewAllPlants() throws PlantException;
+	public List<Plant> viewAllPlants() throws PlantException;
 
-	    public List<Plant> viewPlantsByPlantType(String type) throws PlantException;
-	    
-		public List<Plant> viewPlantsByHeight(Double min,Double max) throws PlantException;
-		
-		public List<Plant> viewPlantsBySpread(String plantSpread) throws PlantException;
-		
-		public List<Plant> viewPlantsByTemperature(Integer min,Integer max) throws PlantException;
-		
-		public List<Plant> viewPlantsByCost(Double min,Double max) throws PlantException;
+	public List<Plant> viewPlantsByPlantType(String type) throws PlantException;
+
+	public List<Plant> viewPlantsByHeight(Double min, Double max) throws PlantException;
+
+	public List<Plant> viewPlantsBySpread(String plantSpread) throws PlantException;
+
+	public List<Plant> viewPlantsByTemperature(Double min, Double max) throws PlantException;
+
+	public List<Plant> viewPlantsByCost(Double min, Double max) throws PlantException;
 }
